// Front-Sistema-Web/src/Models/ModelPersonalizacao.ts
export class Personalizacao {
  cnpj: string = "";
  inscricao_estadual: string = "";
  inscricao_municipal: string = "";
  nome_empresarial: string = "";
  nome_fantasia: string = "";
  data_abertura: string = "";
  rua: string = "";
  numero: string = "";
  complemento: string = "";
  cep: string = "";
  bairro: string = "";
  cidade: string = "";
  uf: string = "";
  telefone: string = "";
  emp_imagem: string = "";
  logomarca_claro: string = "";
  logomarca_escuro: string = "";
  ID_empresa: number = 0;
}